Signal2DTools to Signal2D
