+++ hyperspy/_signals/spectrum.py
@@ -17,9 +17,11 @@
 # along with  HyperSpy.  If not, see <http://www.gnu.org/licenses/>.
 
 from hyperspy._signals.signal1D import Signal1D
+from hyperspy._signals.signal2D import Signal2DTools
 from hyperspy.misc.hspy_warnings import VisibleDeprecationWarning
 
 
-class Spectrum(Signal1D):
+class Spectrum(Signal1D,
+               Signal2DTools,):
     VisibleDeprecationWarning('The Spectrum class will be deprecated from\
                               version 1.0.0 and replaced with Signal1D')
