+++ hyperspy/drawing/image.py
@@ -17,7 +17,6 @@
 # along with  HyperSpy.  If not, see <http://www.gnu.org/licenses/>.
 
 
-
 import math
 import warnings
 
@@ -30,9 +29,6 @@ from hyperspy.drawing import utils
 from hyperspy.gui.tools import ImageContrastEditor
 from hyperspy.misc import math_tools
 from hyperspy.misc import rgb_tools
-from hyperspy.misc.signal2D_tools import (contrast_stretching,
-                                          MPL_DIVERGING_COLORMAPS,
-                                          centre_colormap_values)
 from hyperspy.drawing.figure import BlittedFigure
 
 
@@ -176,7 +172,7 @@ class ImagePlot(BlittedFigure):
             return
         if 'complex' in data.dtype.name:
             data = np.log(np.abs(data))
-        vmin, vmax = contrast_stretching(data, self.saturated_pixels)
+        vmin, vmax = utils.contrast_stretching(data, self.saturated_pixels)
         if self.vmin is None or self.auto_contrast:
             self.vmin = vmin
         if self.vmax is None or self.auto_contrast:
@@ -278,7 +274,7 @@ class ImagePlot(BlittedFigure):
                 cmap = kwargs["cmap"]
             else:
                 cmap = plt.cm.get_cmap().name
-            if cmap in MPL_DIVERGING_COLORMAPS:
+            if cmap in utils.MPL_DIVERGING_COLORMAPS:
                 self.centre_colormap = True
             else:
                 self.centre_colormap = False
@@ -321,7 +317,7 @@ class ImagePlot(BlittedFigure):
         if self.no_nans:
             data = np.nan_to_num(data)
         if self.centre_colormap:
-            vmin, vmax = centre_colormap_values(self.vmin, self.vmax)
+            vmin, vmax = utils.centre_colormap_values(self.vmin, self.vmax)
         else:
             vmin, vmax = self.vmin, self.vmax
         if ims:
