+++ hyperspy/drawing/utils.py
@@ -27,10 +27,76 @@ import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib as mpl
 
-from hyperspy.misc.signal2D_tools import (contrast_stretching,
-                                       MPL_DIVERGING_COLORMAPS,
-                                       centre_colormap_values)
 from hyperspy.defaults_parser import preferences
+
+
+def contrast_stretching(data, saturated_pixels):
+    """Calculate bounds that leaves out a given percentage of the data.
+
+    Parameters
+    ----------
+    data: numpy array
+    saturated_pixels: scalar
+        The percentage of pixels that are left out of the bounds.  For example,
+        the low and high bounds of a value of 1 are the 0.5% and 99.5%
+        percentiles. It must be in the [0, 100] range.
+
+    Returns
+    -------
+    vmin, vmax: scalar
+        The low and high bounds
+
+    Raises
+    ------
+    ValueError if the value of `saturated_pixels` is out of the valid range.
+
+    """
+    # Sanity check
+    if not 0 <= saturated_pixels <= 100:
+        raise ValueError(
+            "saturated_pixels must be a scalar in the range[0, 100]")
+    nans = np.isnan(data)
+    if nans.any():
+        data = data[~nans]
+    vmin = np.percentile(data, saturated_pixels / 2.)
+    vmax = np.percentile(data, 100 - saturated_pixels / 2.)
+    return vmin, vmax
+
+
+MPL_DIVERGING_COLORMAPS = [
+    "BrBG",
+    "bwr",
+    "coolwarm",
+    "PiYG",
+    "PRGn",
+    "PuOr",
+    "RdBu",
+    "RdGy",
+    "RdYIBu",
+    "RdYIGn",
+    "seismic",
+    "Spectral", ]
+# Add reversed colormaps
+MPL_DIVERGING_COLORMAPS += [cmap + "_r" for cmap in MPL_DIVERGING_COLORMAPS]
+
+
+def centre_colormap_values(vmin, vmax):
+    """Calculate vmin and vmax to set the colormap midpoint to zero.
+
+    Parameters
+    ----------
+    vmin, vmax : scalar
+        The range of data to display.
+
+    Returns
+    -------
+    cvmin, cvmax : scalar
+        The values to obtain a centre colormap.
+
+    """
+
+    absmax = max(abs(vmin), abs(vmax))
+    return -absmax, absmax
 
 
 def create_figure(window_title=None,
