+++ hyperspy/tests/misc/test_image_tools.py
@@ -1,6 +1,6 @@
 import nose.tools
 import numpy as np
-from hyperspy.misc.signal2D_tools import contrast_stretching
+from hyperspy.drawing.utils import contrast_stretching
 
 
 class TestImageStretching:
